refactor: Rename `lib/resolveOutputs.js`
