+++ lib/output-command.js
@@ -5,7 +5,7 @@ const { writeText } = require('@serverless/utils/log');
 const log = require('./log');
 const isAuthenticated = require('./is-authenticated');
 const throwAuthError = require('./throwAuthError');
-const resolveOutputs = require('./resolveOutputs');
+const resolveOutputs = require('./resolve-outputs');
 const resolveOutput = require('./resolve-output');
 
 const resolveInput = function (ctx) {
