+++ lib/resolve-outputs.js
